Add Routing for print map
