+++ app/app-routing.module.ts
@@ -3,12 +3,15 @@ import { AppComponent } from './app.component';
 import { TouchpadComponent } from './touchpad/touchpad.component'
 import { EditorComponent } from './editor/editor.component'
 import { RouterModule, Routes } from '@angular/router';
+import { PrintMapComponent } from "./printable-map/PrintMap.component";
+
 
 /* Routing Module */
 const appRoutes: Routes = [
   { path: '', redirectTo: '/editor', pathMatch: 'full'},
   { path: 'editor', component: EditorComponent },
-  { path: 'touchpad-voice', component: TouchpadComponent },
+  { path: 'touchpad-voice', component: TouchpadComponent }
+  ,{ path: 'printable-map/:id', component: PrintMapComponent }
 ];
 
 @NgModule({
@@ -16,6 +19,7 @@ const appRoutes: Routes = [
     RouterModule.forRoot(appRoutes)
   ],
   exports:    [ RouterModule ]
+
 })
 
 export class AppRoutingModule {}
